+++ lass_audio/apex/tests/L0/run_mixed_adam/test_fp16_optimizer.py
@@ -14,9 +14,9 @@ class TestFP16Optimizer(unittest.TestCase):
         self.N = N
         self.D_in = D_in
         self.D_out = D_out
-        self.x = torch.randn((N, D_in), dtype=torch.float16, device='cuda')
-        self.ref_model = torch.nn.Linear(D_in, D_out).half()
-        self.tst_model = torch.nn.Linear(D_in, D_out).half()
+        self.x = torch.randn((N, D_in), dtype=torch.float32, device='cuda')
+        self.ref_model = torch.nn.Linear(D_in, D_out)
+        self.tst_model = torch.nn.Linear(D_in, D_out)
         for p, q in zip(self.tst_model.parameters(), self.ref_model.parameters()):
             p.data.copy_(q.data)
